weather: little refactor
Sometimes the weather was empty due to api error and it appended it to
the file. This fix ensures the response is never empty and if it is it
does not append such result to the cache file. Instead we exit and wait
for another time to call for the weather api.